+++ ext/stub.c
@@ -221,6 +221,7 @@ zend_class_entry *stub_issue2656_ce;
 zend_class_entry *stub_issue2666_ce;
 zend_class_entry *stub_issue2674_ce;
 zend_class_entry *stub_issue2679_ce;
+zend_class_entry *stub_issue2682_ce;
 zend_class_entry *stub_issue663_ce;
 zend_class_entry *stub_issue760_ce;
 zend_class_entry *stub_issue808_ce;
@@ -523,6 +524,7 @@ static PHP_MINIT_FUNCTION(stub)
 	ZEPHIR_INIT(Stub_Issue2666);
 	ZEPHIR_INIT(Stub_Issue2674);
 	ZEPHIR_INIT(Stub_Issue2679);
+	ZEPHIR_INIT(Stub_Issue2682);
 	ZEPHIR_INIT(Stub_Issue663);
 	ZEPHIR_INIT(Stub_Issue760);
 	ZEPHIR_INIT(Stub_Issue808);