+++ ext/stub/quantum.zep.c
@@ -34,11 +34,11 @@ ZEPHIR_INIT_CLASS(Stub_Quantum)
 PHP_METHOD(Stub_Quantum, harmos)
 {
 	zval _8$$4, _10$$4, _11$$4;
-	zend_bool _12, _15$$5, _34$$5, _50$$5;
+	zend_bool _12, _15$$5, _35$$5, _52$$5;
 	zephir_method_globals *ZEPHIR_METHOD_GLOBALS_PTR = NULL;
-	zephir_fcall_cache_entry *_5 = NULL, *_55 = NULL;
-	zend_long ZEPHIR_LAST_CALL_STATUS, i = 0, j = 0, n = 0, _13, _14, _16$$5, _17$$5, _35$$5, _36$$5;
-	zval *x_param = NULL, psr, psi, p2, v, paramater, fp, tmp, _0, _1, _2$$4, _3$$4, _4$$4, _6$$4, _7$$4, _9$$4, _18$$6, _19$$6, _20$$6, _21$$6, _22$$6, _23$$6, _24$$6, _25$$6, _26$$6, _27$$6, _28$$6, _29$$6, _30$$6, _31$$6, _32$$6, _33$$6, _37$$7, _38$$7, _39$$7, _40$$7, _41$$7, _42$$7, _43$$7, _44$$7, _45$$7, _46$$7, _47$$7, _48$$7, _49$$7, _51$$9, _52$$9, _53$$9, _54$$9, _56$$8, _57$$10, _58$$10, _59$$10;
+	zephir_fcall_cache_entry *_5 = NULL, *_57 = NULL;
+	zend_long ZEPHIR_LAST_CALL_STATUS, i = 0, j = 0, n = 0, _13, _14, _16$$5, _17$$5, _36$$5, _37$$5;
+	zval *x_param = NULL, psr, psi, p2, v, paramater, fp, tmp, _0, _1, _2$$4, _3$$4, _4$$4, _6$$4, _7$$4, _9$$4, _18$$6, _19$$6, _20$$6, _21$$6, _22$$6, _23$$6, _24$$6, _25$$6, _26$$6, _27$$6, _28$$6, _29$$6, _30$$6, _31$$6, _32$$6, _33$$6, _34$$6, _38$$7, _39$$7, _40$$7, _41$$7, _42$$7, _43$$7, _44$$7, _45$$7, _46$$7, _47$$7, _48$$7, _49$$7, _50$$7, _51$$7, _53$$9, _54$$9, _55$$9, _56$$9, _58$$8, _59$$10, _60$$10, _61$$10;
 	double x, dt = 0, dx = 0, k0 = 0, item_psr = 0, item_psi = 0;
 
 	ZVAL_UNDEF(&psr);
@@ -72,7 +72,7 @@ PHP_METHOD(Stub_Quantum, harmos)
 	ZVAL_UNDEF(&_31$$6);
 	ZVAL_UNDEF(&_32$$6);
 	ZVAL_UNDEF(&_33$$6);
-	ZVAL_UNDEF(&_37$$7);
+	ZVAL_UNDEF(&_34$$6);
 	ZVAL_UNDEF(&_38$$7);
 	ZVAL_UNDEF(&_39$$7);
 	ZVAL_UNDEF(&_40$$7);
@@ -85,14 +85,16 @@ PHP_METHOD(Stub_Quantum, harmos)
 	ZVAL_UNDEF(&_47$$7);
 	ZVAL_UNDEF(&_48$$7);
 	ZVAL_UNDEF(&_49$$7);
-	ZVAL_UNDEF(&_51$$9);
-	ZVAL_UNDEF(&_52$$9);
+	ZVAL_UNDEF(&_50$$7);
+	ZVAL_UNDEF(&_51$$7);
 	ZVAL_UNDEF(&_53$$9);
 	ZVAL_UNDEF(&_54$$9);
-	ZVAL_UNDEF(&_56$$8);
-	ZVAL_UNDEF(&_57$$10);
-	ZVAL_UNDEF(&_58$$10);
+	ZVAL_UNDEF(&_55$$9);
+	ZVAL_UNDEF(&_56$$9);
+	ZVAL_UNDEF(&_58$$8);
 	ZVAL_UNDEF(&_59$$10);
+	ZVAL_UNDEF(&_60$$10);
+	ZVAL_UNDEF(&_61$$10);
 	ZVAL_UNDEF(&_8$$4);
 	ZVAL_UNDEF(&_10$$4);
 	ZVAL_UNDEF(&_11$$4);
@@ -209,142 +211,158 @@ PHP_METHOD(Stub_Quantum, harmos)
 					}
 					i = _16$$5;
 					zephir_array_fetch_long(&_18$$6, &psr, i, PH_NOISY | PH_READONLY, "stub/quantum.zep", 50);
-					zephir_array_fetch_long(&_19$$6, &_18$$6, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 50);
+					ZEPHIR_OBS_NVAR(&_19$$6);
+					zephir_array_fetch_long(&_19$$6, &_18$$6, 0, PH_NOISY, "stub/quantum.zep", 50);
 					zephir_array_fetch_long(&_20$$6, &paramater, 2, PH_NOISY | PH_READONLY, "stub/quantum.zep", 50);
 					zephir_array_fetch_long(&_21$$6, &psi, (i + 1), PH_NOISY | PH_READONLY, "stub/quantum.zep", 50);
-					zephir_array_fetch_long(&_22$$6, &_21$$6, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 50);
+					ZEPHIR_OBS_NVAR(&_22$$6);
+					zephir_array_fetch_long(&_22$$6, &_21$$6, 0, PH_NOISY, "stub/quantum.zep", 50);
 					zephir_array_fetch_long(&_23$$6, &psi, (i - 1), PH_NOISY | PH_READONLY, "stub/quantum.zep", 50);
-					zephir_array_fetch_long(&_24$$6, &_23$$6, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
+					ZEPHIR_OBS_NVAR(&_24$$6);
+					zephir_array_fetch_long(&_24$$6, &_23$$6, 0, PH_NOISY, "stub/quantum.zep", 51);
 					ZEPHIR_INIT_NVAR(&_25$$6);
 					zephir_add_function(&_25$$6, &_22$$6, &_24$$6);
-					zephir_array_fetch_long(&_22$$6, &tmp, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
-					zephir_array_fetch_long(&_24$$6, &psi, i, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
-					zephir_array_fetch_long(&_26$$6, &_24$$6, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
-					ZEPHIR_INIT_NVAR(&_27$$6);
-					mul_function(&_27$$6, &_22$$6, &_26$$6);
+					zephir_array_fetch_long(&_26$$6, &tmp, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
+					zephir_array_fetch_long(&_27$$6, &psi, i, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
+					ZEPHIR_OBS_NVAR(&_22$$6);
+					zephir_array_fetch_long(&_22$$6, &_27$$6, 0, PH_NOISY, "stub/quantum.zep", 51);
 					ZEPHIR_INIT_NVAR(&_28$$6);
-					zephir_sub_function(&_28$$6, &_25$$6, &_27$$6);
-					ZEPHIR_INIT_NVAR(&_25$$6);
-					mul_function(&_25$$6, &_20$$6, &_28$$6);
-					zephir_array_fetch_long(&_20$$6, &paramater, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
-					zephir_array_fetch_long(&_22$$6, &paramater, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
-					ZEPHIR_INIT_NVAR(&_28$$6);
-					mul_function(&_28$$6, &_20$$6, &_22$$6);
-					ZEPHIR_INIT_NVAR(&_27$$6);
-					div_function(&_27$$6, &_25$$6, &_28$$6);
-					ZEPHIR_INIT_NVAR(&_25$$6);
-					zephir_sub_function(&_25$$6, &_19$$6, &_27$$6);
-					zephir_array_fetch_long(&_19$$6, &paramater, 2, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
-					zephir_array_fetch_long(&_20$$6, &v, i, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
-					zephir_array_fetch_long(&_22$$6, &_20$$6, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
-					ZEPHIR_INIT_NVAR(&_27$$6);
-					mul_function(&_27$$6, &_19$$6, &_22$$6);
-					zephir_array_fetch_long(&_19$$6, &psi, i, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
-					zephir_array_fetch_long(&_22$$6, &_19$$6, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
+					mul_function(&_28$$6, &_26$$6, &_22$$6);
 					ZEPHIR_INIT_NVAR(&_29$$6);
-					mul_function(&_29$$6, &_27$$6, &_22$$6);
-					ZEPHIR_INIT_NVAR(&_28$$6);
-					zephir_add_function(&_28$$6, &_25$$6, &_29$$6);
-					zephir_array_update_multi(&psr, &_28$$6, SL("ll"), 2, (zend_long) i, (zend_long) 1);
-					zephir_array_fetch_long(&_22$$6, &psr, i, PH_NOISY | PH_READONLY, "stub/quantum.zep", 53);
-					zephir_array_fetch_long(&_26$$6, &_22$$6, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 53);
-					zephir_array_fetch_long(&_30$$6, &psr, i, PH_NOISY | PH_READONLY, "stub/quantum.zep", 53);
-					zephir_array_fetch_long(&_31$$6, &_30$$6, 1, PH_NOISY | PH_READONLY, "stub/quantum.zep", 53);
+					zephir_sub_function(&_29$$6, &_25$$6, &_28$$6);
 					ZEPHIR_INIT_NVAR(&_25$$6);
-					mul_function(&_25$$6, &_26$$6, &_31$$6);
-					zephir_array_fetch_long(&_26$$6, &psi, i, PH_NOISY | PH_READONLY, "stub/quantum.zep", 53);
-					zephir_array_fetch_long(&_31$$6, &_26$$6, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 53);
-					zephir_array_fetch_long(&_32$$6, &psi, i, PH_NOISY | PH_READONLY, "stub/quantum.zep", 53);
-					zephir_array_fetch_long(&_33$$6, &_32$$6, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 53);
-					ZEPHIR_INIT_NVAR(&_27$$6);
-					mul_function(&_27$$6, &_31$$6, &_33$$6);
+					mul_function(&_25$$6, &_20$$6, &_29$$6);
+					zephir_array_fetch_long(&_20$$6, &paramater, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
+					zephir_array_fetch_long(&_26$$6, &paramater, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
+					ZEPHIR_INIT_NVAR(&_29$$6);
+					mul_function(&_29$$6, &_20$$6, &_26$$6);
 					ZEPHIR_INIT_NVAR(&_28$$6);
-					zephir_add_function(&_28$$6, &_25$$6, &_27$$6);
-					zephir_array_update_long(&p2, i, &_28$$6, PH_COPY | PH_SEPARATE ZEPHIR_DEBUG_PARAMS_DUMMY);
+					div_function(&_28$$6, &_25$$6, &_29$$6);
+					ZEPHIR_INIT_NVAR(&_25$$6);
+					zephir_sub_function(&_25$$6, &_19$$6, &_28$$6);
+					zephir_array_fetch_long(&_20$$6, &paramater, 2, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
+					zephir_array_fetch_long(&_26$$6, &v, i, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
+					ZEPHIR_OBS_NVAR(&_19$$6);
+					zephir_array_fetch_long(&_19$$6, &_26$$6, 0, PH_NOISY, "stub/quantum.zep", 51);
+					ZEPHIR_INIT_NVAR(&_28$$6);
+					mul_function(&_28$$6, &_20$$6, &_19$$6);
+					zephir_array_fetch_long(&_20$$6, &psi, i, PH_NOISY | PH_READONLY, "stub/quantum.zep", 51);
+					ZEPHIR_OBS_NVAR(&_19$$6);
+					zephir_array_fetch_long(&_19$$6, &_20$$6, 0, PH_NOISY, "stub/quantum.zep", 51);
+					ZEPHIR_INIT_NVAR(&_30$$6);
+					mul_function(&_30$$6, &_28$$6, &_19$$6);
+					ZEPHIR_INIT_NVAR(&_29$$6);
+					zephir_add_function(&_29$$6, &_25$$6, &_30$$6);
+					zephir_array_update_multi(&psr, &_29$$6, SL("ll"), 2, (zend_long) i, (zend_long) 1);
+					zephir_array_fetch_long(&_31$$6, &psr, i, PH_NOISY | PH_READONLY, "stub/quantum.zep", 53);
+					ZEPHIR_OBS_NVAR(&_19$$6);
+					zephir_array_fetch_long(&_19$$6, &_31$$6, 0, PH_NOISY, "stub/quantum.zep", 53);
+					zephir_array_fetch_long(&_32$$6, &psr, i, PH_NOISY | PH_READONLY, "stub/quantum.zep", 53);
+					ZEPHIR_OBS_NVAR(&_22$$6);
+					zephir_array_fetch_long(&_22$$6, &_32$$6, 1, PH_NOISY, "stub/quantum.zep", 53);
+					ZEPHIR_INIT_NVAR(&_25$$6);
+					mul_function(&_25$$6, &_19$$6, &_22$$6);
+					zephir_array_fetch_long(&_33$$6, &psi, i, PH_NOISY | PH_READONLY, "stub/quantum.zep", 53);
+					ZEPHIR_OBS_NVAR(&_19$$6);
+					zephir_array_fetch_long(&_19$$6, &_33$$6, 0, PH_NOISY, "stub/quantum.zep", 53);
+					zephir_array_fetch_long(&_34$$6, &psi, i, PH_NOISY | PH_READONLY, "stub/quantum.zep", 53);
+					ZEPHIR_OBS_NVAR(&_22$$6);
+					zephir_array_fetch_long(&_22$$6, &_34$$6, 0, PH_NOISY, "stub/quantum.zep", 53);
+					ZEPHIR_INIT_NVAR(&_28$$6);
+					mul_function(&_28$$6, &_19$$6, &_22$$6);
+					ZEPHIR_INIT_NVAR(&_29$$6);
+					zephir_add_function(&_29$$6, &_25$$6, &_28$$6);
+					zephir_array_update_long(&p2, i, &_29$$6, PH_COPY | PH_SEPARATE ZEPHIR_DEBUG_PARAMS_DUMMY);
 				}
 			}
-			_36$$5 = (751 - 1);
-			_35$$5 = 1;
-			_34$$5 = 0;
-			if (_35$$5 <= _36$$5) {
+			_37$$5 = (751 - 1);
+			_36$$5 = 1;
+			_35$$5 = 0;
+			if (_36$$5 <= _37$$5) {
 				while (1) {
-					if (_34$$5) {
-						_35$$5++;
-						if (!(_35$$5 <= _36$$5)) {
+					if (_35$$5) {
+						_36$$5++;
+						if (!(_36$$5 <= _37$$5)) {
 							break;
 						}
 					} else {
-						_34$$5 = 1;
+						_35$$5 = 1;
 					}
-					j = _35$$5;
-					ZEPHIR_INIT_NVAR(&_37$$7);
-					ZVAL_LONG(&_37$$7, 0);
-					zephir_array_update_multi(&psr, &_37$$7, SL("ll"), 2, (zend_long) 0, (zend_long) 1);
-					ZEPHIR_INIT_NVAR(&_37$$7);
-					ZVAL_LONG(&_37$$7, 0);
-					zephir_array_update_multi(&psr, &_37$$7, SL("ll"), 2, (zend_long) 751, (zend_long) 1);
-					zephir_array_fetch_long(&_38$$7, &psi, j, PH_NOISY | PH_READONLY, "stub/quantum.zep", 58);
-					zephir_array_fetch_long(&_39$$7, &_38$$7, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 58);
-					zephir_array_fetch_long(&_40$$7, &paramater, 2, PH_NOISY | PH_READONLY, "stub/quantum.zep", 58);
-					zephir_array_fetch_long(&_41$$7, &psr, (j + 1), PH_NOISY | PH_READONLY, "stub/quantum.zep", 58);
-					zephir_array_fetch_long(&_42$$7, &_41$$7, 1, PH_NOISY | PH_READONLY, "stub/quantum.zep", 58);
-					zephir_array_fetch_long(&_43$$7, &psr, (j - 1), PH_NOISY | PH_READONLY, "stub/quantum.zep", 58);
-					zephir_array_fetch_long(&_44$$7, &_43$$7, 1, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
-					ZEPHIR_INIT_NVAR(&_45$$7);
-					zephir_add_function(&_45$$7, &_42$$7, &_44$$7);
-					zephir_array_fetch_long(&_42$$7, &tmp, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
-					zephir_array_fetch_long(&_44$$7, &psr, j, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
-					zephir_array_fetch_long(&_46$$7, &_44$$7, 1, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
-					ZEPHIR_INIT_NVAR(&_47$$7);
-					mul_function(&_47$$7, &_42$$7, &_46$$7);
-					ZEPHIR_INIT_NVAR(&_48$$7);
-					zephir_sub_function(&_48$$7, &_45$$7, &_47$$7);
-					ZEPHIR_INIT_NVAR(&_45$$7);
-					mul_function(&_45$$7, &_40$$7, &_48$$7);
-					zephir_array_fetch_long(&_40$$7, &paramater, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
-					zephir_array_fetch_long(&_42$$7, &paramater, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
-					ZEPHIR_INIT_NVAR(&_48$$7);
-					mul_function(&_48$$7, &_40$$7, &_42$$7);
-					ZEPHIR_INIT_NVAR(&_47$$7);
-					div_function(&_47$$7, &_45$$7, &_48$$7);
-					ZEPHIR_INIT_NVAR(&_45$$7);
-					zephir_add_function(&_45$$7, &_39$$7, &_47$$7);
-					zephir_array_fetch_long(&_39$$7, &paramater, 2, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
-					zephir_array_fetch_long(&_40$$7, &v, j, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
-					zephir_array_fetch_long(&_42$$7, &_40$$7, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
-					ZEPHIR_INIT_NVAR(&_47$$7);
-					mul_function(&_47$$7, &_39$$7, &_42$$7);
-					zephir_array_fetch_long(&_39$$7, &psr, j, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
-					zephir_array_fetch_long(&_42$$7, &_39$$7, 1, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
+					j = _36$$5;
+					ZEPHIR_INIT_NVAR(&_38$$7);
+					ZVAL_LONG(&_38$$7, 0);
+					zephir_array_update_multi(&psr, &_38$$7, SL("ll"), 2, (zend_long) 0, (zend_long) 1);
+					ZEPHIR_INIT_NVAR(&_38$$7);
+					ZVAL_LONG(&_38$$7, 0);
+					zephir_array_update_multi(&psr, &_38$$7, SL("ll"), 2, (zend_long) 751, (zend_long) 1);
+					zephir_array_fetch_long(&_39$$7, &psi, j, PH_NOISY | PH_READONLY, "stub/quantum.zep", 58);
+					ZEPHIR_OBS_NVAR(&_40$$7);
+					zephir_array_fetch_long(&_40$$7, &_39$$7, 0, PH_NOISY, "stub/quantum.zep", 58);
+					zephir_array_fetch_long(&_41$$7, &paramater, 2, PH_NOISY | PH_READONLY, "stub/quantum.zep", 58);
+					zephir_array_fetch_long(&_42$$7, &psr, (j + 1), PH_NOISY | PH_READONLY, "stub/quantum.zep", 58);
+					ZEPHIR_OBS_NVAR(&_43$$7);
+					zephir_array_fetch_long(&_43$$7, &_42$$7, 1, PH_NOISY, "stub/quantum.zep", 58);
+					zephir_array_fetch_long(&_44$$7, &psr, (j - 1), PH_NOISY | PH_READONLY, "stub/quantum.zep", 58);
+					ZEPHIR_OBS_NVAR(&_45$$7);
+					zephir_array_fetch_long(&_45$$7, &_44$$7, 1, PH_NOISY, "stub/quantum.zep", 59);
+					ZEPHIR_INIT_NVAR(&_46$$7);
+					zephir_add_function(&_46$$7, &_43$$7, &_45$$7);
+					zephir_array_fetch_long(&_47$$7, &tmp, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
+					zephir_array_fetch_long(&_48$$7, &psr, j, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
+					ZEPHIR_OBS_NVAR(&_43$$7);
+					zephir_array_fetch_long(&_43$$7, &_48$$7, 1, PH_NOISY, "stub/quantum.zep", 59);
 					ZEPHIR_INIT_NVAR(&_49$$7);
-					mul_function(&_49$$7, &_47$$7, &_42$$7);
-					ZEPHIR_INIT_NVAR(&_48$$7);
-					zephir_sub_function(&_48$$7, &_45$$7, &_49$$7);
-					zephir_array_update_multi(&psi, &_48$$7, SL("ll"), 2, (zend_long) j, (zend_long) 1);
+					mul_function(&_49$$7, &_47$$7, &_43$$7);
+					ZEPHIR_INIT_NVAR(&_50$$7);
+					zephir_sub_function(&_50$$7, &_46$$7, &_49$$7);
+					ZEPHIR_INIT_NVAR(&_46$$7);
+					mul_function(&_46$$7, &_41$$7, &_50$$7);
+					zephir_array_fetch_long(&_41$$7, &paramater, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
+					zephir_array_fetch_long(&_47$$7, &paramater, 0, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
+					ZEPHIR_INIT_NVAR(&_50$$7);
+					mul_function(&_50$$7, &_41$$7, &_47$$7);
+					ZEPHIR_INIT_NVAR(&_49$$7);
+					div_function(&_49$$7, &_46$$7, &_50$$7);
+					ZEPHIR_INIT_NVAR(&_46$$7);
+					zephir_add_function(&_46$$7, &_40$$7, &_49$$7);
+					zephir_array_fetch_long(&_41$$7, &paramater, 2, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
+					zephir_array_fetch_long(&_47$$7, &v, j, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
+					ZEPHIR_OBS_NVAR(&_40$$7);
+					zephir_array_fetch_long(&_40$$7, &_47$$7, 0, PH_NOISY, "stub/quantum.zep", 59);
+					ZEPHIR_INIT_NVAR(&_49$$7);
+					mul_function(&_49$$7, &_41$$7, &_40$$7);
+					zephir_array_fetch_long(&_41$$7, &psr, j, PH_NOISY | PH_READONLY, "stub/quantum.zep", 59);
+					ZEPHIR_OBS_NVAR(&_40$$7);
+					zephir_array_fetch_long(&_40$$7, &_41$$7, 1, PH_NOISY, "stub/quantum.zep", 59);
+					ZEPHIR_INIT_NVAR(&_51$$7);
+					mul_function(&_51$$7, &_49$$7, &_40$$7);
+					ZEPHIR_INIT_NVAR(&_50$$7);
+					zephir_sub_function(&_50$$7, &_46$$7, &_51$$7);
+					zephir_array_update_multi(&psi, &_50$$7, SL("ll"), 2, (zend_long) j, (zend_long) 1);
 				}
 			}
-			_50$$5 = n == 0;
-			if (!(_50$$5)) {
-				_50$$5 = zephir_safe_mod_long_long(n, 2000) == 0;
+			_52$$5 = n == 0;
+			if (!(_52$$5)) {
+				_52$$5 = zephir_safe_mod_long_long(n, 2000) == 0;
 			}
-			if (_50$$5) {
+			if (_52$$5) {
 				i = 1;
 				while (1) {
 					if (!(i < (751 - 1))) {
 						break;
 					}
-					zephir_array_fetch_long(&_51$$9, &p2, i, PH_NOISY | PH_READONLY, "stub/quantum.zep", 67);
-					ZEPHIR_INIT_NVAR(&_52$$9);
-					ZVAL_STRING(&_52$$9, "%16.8lf %16.8lf %16.8lf \n");
-					ZVAL_DOUBLE(&_53$$9, ((double) i * dx));
-					ZVAL_DOUBLE(&_54$$9, ((double) n * dt));
-					ZEPHIR_CALL_FUNCTION(NULL, "fprintf", &_55, 94, &fp, &_52$$9, &_53$$9, &_54$$9, &_51$$9);
+					zephir_array_fetch_long(&_53$$9, &p2, i, PH_NOISY | PH_READONLY, "stub/quantum.zep", 67);
+					ZEPHIR_INIT_NVAR(&_54$$9);
+					ZVAL_STRING(&_54$$9, "%16.8lf %16.8lf %16.8lf \n");
+					ZVAL_DOUBLE(&_55$$9, ((double) i * dx));
+					ZVAL_DOUBLE(&_56$$9, ((double) n * dt));
+					ZEPHIR_CALL_FUNCTION(NULL, "fprintf", &_57, 94, &fp, &_54$$9, &_55$$9, &_56$$9, &_53$$9);
 					zephir_check_call_status();
 					i = (i + 10);
 				}
-				ZEPHIR_INIT_NVAR(&_56$$8);
-				ZVAL_STRING(&_56$$8, "\n");
-				ZEPHIR_CALL_FUNCTION(NULL, "fprintf", &_55, 94, &fp, &_56$$8);
+				ZEPHIR_INIT_NVAR(&_58$$8);
+				ZVAL_STRING(&_58$$8, "\n");
+				ZEPHIR_CALL_FUNCTION(NULL, "fprintf", &_57, 94, &fp, &_58$$8);
 				zephir_check_call_status();
 			}
 			j = 1;
@@ -352,12 +370,14 @@ PHP_METHOD(Stub_Quantum, harmos)
 				if (!(j < (751 - 1))) {
 					break;
 				}
-				zephir_array_fetch_long(&_57$$10, &psi, j, PH_NOISY | PH_READONLY, "stub/quantum.zep", 76);
-				zephir_array_fetch_long(&_58$$10, &_57$$10, 1, PH_NOISY | PH_READONLY, "stub/quantum.zep", 76);
-				zephir_array_update_multi(&psi, &_58$$10, SL("ll"), 2, (zend_long) j, (zend_long) 0);
-				zephir_array_fetch_long(&_58$$10, &psr, j, PH_NOISY | PH_READONLY, "stub/quantum.zep", 77);
-				zephir_array_fetch_long(&_59$$10, &_58$$10, 1, PH_NOISY | PH_READONLY, "stub/quantum.zep", 77);
-				zephir_array_update_multi(&psr, &_59$$10, SL("ll"), 2, (zend_long) j, (zend_long) 0);
+				zephir_array_fetch_long(&_59$$10, &psi, j, PH_NOISY | PH_READONLY, "stub/quantum.zep", 76);
+				ZEPHIR_OBS_NVAR(&_60$$10);
+				zephir_array_fetch_long(&_60$$10, &_59$$10, 1, PH_NOISY, "stub/quantum.zep", 76);
+				zephir_array_update_multi(&psi, &_60$$10, SL("ll"), 2, (zend_long) j, (zend_long) 0);
+				zephir_array_fetch_long(&_61$$10, &psr, j, PH_NOISY | PH_READONLY, "stub/quantum.zep", 77);
+				ZEPHIR_OBS_NVAR(&_60$$10);
+				zephir_array_fetch_long(&_60$$10, &_61$$10, 1, PH_NOISY, "stub/quantum.zep", 77);
+				zephir_array_update_multi(&psr, &_60$$10, SL("ll"), 2, (zend_long) j, (zend_long) 0);
 				j++;
 			}
 		}